Move check_for_models to sunspec.core.util
