+++ sunspec/core/test/test_device.py
@@ -35,7 +35,7 @@ import sunspec.core.util as util
 import sunspec.core.pics as pics
 import sunspec.core.smdx as smdx
 import sunspec.core.suns as suns
-import sunspec.core.test.util
+import sunspec.core.util
 
 test_device_pointtype_smdx_1 = """
 <sunSpecModels v="1">
@@ -304,7 +304,7 @@ class TestDevice(unittest.TestCase):
                                        os.path.join(os.path.dirname(path),
                                                     'devices')])
 
-        sunspec.core.test.util.check_for_models(pathlist=self.pathlist)
+        sunspec.core.util.check_for_models(pathlist=self.pathlist)
 
     def test_device_modeltype(self):
         mt = device.ModelType()
