+++ sunspec/core/util.py
@@ -26,6 +26,9 @@ import struct
 import zipfile
 import array
 
+import sunspec.core.device
+
+
 class SunSpecError(Exception):
     pass
 
@@ -248,3 +251,12 @@ class PathList(object):
 
 
 
+def check_for_models(pathlist):
+    # The common model (1) should be accessible.
+    try:
+        sunspec.core.device.model_type_get(1)
+    except:
+        raise Exception(
+            'Unable to open common model (1).  '
+            'Make certain model repository is available.'
+        )
